requirements.txt
aiohttp==3.7.4
fastapi==0.63.0
pytest==6.2.4
pytest_asyncio==0.15.1
tortoise-orm==0.17.1
uvicorn==0.13.4
